fix(R5-5): use Long literal in Window.toBoundary instead of truncating to Int
Previously, frame boundary values were cast via `.toInt`, silently
truncating values outside Int range. Now passes the Long value directly
to `Column.lit(value)`, producing a correct Long literal Expression.

+++ src/main/scala/org/apache/spark/sql/Column.scala
@@ -594,6 +594,6 @@ object Window:
       Expression.Window.WindowFrame.FrameBoundary.newBuilder()
         .setCurrentRow(true).build()
     else
-      val litExpr = Column.lit(value.toInt).expr
+      val litExpr = Column.lit(value).expr
       Expression.Window.WindowFrame.FrameBoundary.newBuilder()
         .setValue(litExpr).build()
